- add undone tasks to read me - add user login and jwt authentication - create controllers
