+++ .eslintrc.js
@@ -2,5 +2,6 @@ module.exports = {
     "extends": "airbnb-base",
     "rules": {
         "no-underscore-dangle": 0,
+        "consistent-return": 0,
     }
 };
