+++ README.md
@@ -9,5 +9,16 @@ An API for managing recipes that one enters as well as finding recipes created b
   - MongoDB
   - ESLint (AirBnB Style)
 
+
+## Tasks Yet to Be Done:
+  - Tests
+  - Users can add/edit image
+  - Edit image for recipe
+  - Search for recipe
+  - Store recipe to my book
+  - make recipes private or public
+  - Rate a recipe
+  - Data validation
+
 ### Made By
 Tapiwa Lason Mapundematsva
