+++ package.json
@@ -27,6 +27,7 @@
     "debug": "4.1.1",
     "dotenv": "6.2.0",
     "express": "4.16.4",
+    "jsonwebtoken": "8.4.0",
     "mongoose": "5.4.2",
     "morgan": "1.9.1",
     "multer": "1.4.1",
